requirements.txt
# Core Framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0

# Memory & Vector Store
mem0ai>=0.1.0
chromadb>=0.5.0
sentence-transformers>=2.3.0

# Model Providers
anthropic>=0.18.0
openai>=1.12.0
ollama>=0.1.0

# Integrations
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# MCP (Model Context Protocol)
mcp>=1.0.0
httpx>=0.27.0
anyio>=4.3.0

# UI & Utils
streamlit==1.31.0
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Development
pytest==8.0.0
black==24.1.1
ruff==0.2.0
mypy==1.8.0
